finalproject/1/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pyodbc
requests
azure.storage.blob
azure.cosmos
opencv-contrib-python
numpy
pyzbar
